Update build.gradle, rename package
build.gradle points to old class name, updated to
Marvin. Renamed package to Marvin.
