+++ build.gradle
@@ -28,11 +28,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("marvin.Marvin")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "marvin"
     archiveClassifier = null
 }
