+++ src/main/java/marvin/StorageHandler.java
@@ -1,6 +1,6 @@
-package duke;
+package marvin;
 
-import duke.task.TaskList;
+import marvin.task.TaskList;
 
 import java.io.*;
