+++ src/main/java/marvin/command/Command.java
@@ -1,5 +1,5 @@
-package duke.command;
-import duke.task.TaskList;
+package marvin.command;
+import marvin.task.TaskList;
 
 public abstract class Command {
     public abstract void execute(TaskList taskList);
