+++ src/main/java/marvin/task/Deadline.java
@@ -1,4 +1,4 @@
-package duke.task;
+package marvin.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
