+++ src/main/java/marvin/task/Task.java
@@ -1,4 +1,4 @@
-package duke.task;
+package marvin.task;
 
 import java.io.Serializable;
