+++ src/main/java/marvin/task/TaskList.java
@@ -1,4 +1,4 @@
-package duke.task;
+package marvin.task;
 
 import java.io.Serializable;
 import java.util.ArrayList;
