+++ src/main/java/marvin/task/Todo.java
@@ -1,4 +1,4 @@
-package duke.task;
+package marvin.task;
 
 public class Todo extends Task {
     public Todo(String description) {
